recenter system and update text
witch the camera being so far out there was a lot of persepctive being
applied to the text (which was rendered at z = 0). I recentered the
scene to be at (0,200,-700) so that the cameraeye could start closer to
origin. in reotrspect I could have just moved the text back towards the
camera but I like how everything is more relative to the origin this
way.